backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
apscheduler==3.10.4
python-dotenv==1.0.1
python-dateutil==2.9.0
pytest==8.3.4
pytest-asyncio==0.24.0
aiosqlite==0.20.0
